# mvcc: don't allocate keys when computing Revisions