~6% improvement

Before:
```
$ go test -bench=RangeKey100 -run=RangeKey100 -v  -count=4
BenchmarkStoreRangeKey100-8        10000            118881 ns/op           46240 B/op        723 allocs/op
BenchmarkStoreRangeKey100-8        10000            117498 ns/op           46241 B/op        724 allocs/op
BenchmarkStoreRangeKey100-8        10000            119148 ns/op           46241 B/op        724 allocs/op
BenchmarkStoreRangeKey100-8        10000            116682 ns/op           46241 B/op        724 allocs/op
```

After:
```
BenchmarkStoreRangeKey100-8        10000            111767 ns/op           40113 B/op        716 allocs/op
BenchmarkStoreRangeKey100-8        10000            110846 ns/op           40112 B/op        715 allocs/op
BenchmarkStoreRangeKey100-8        10000            111666 ns/op           40112 B/op        715 allocs/op
BenchmarkStoreRangeKey100-8        10000            111672 ns/op           40113 B/op        716 allocs/op
```